requirements.txt
langgraph
langchain
langchain_core
langchain_openai
langchain_groq
langchain-google-genai
pandas
tabulate
python-dotenv
requests
psycopg2-binary
sentence-transformers
faiss-cpu
sqlglot
fastapi
uvicorn
apscheduler
streamlit
# Blockchain
web3
# # Rate Limiting (Redis branch)
# slowapi
# redis
